feat - 사다리 게임 검증 코드 추가 + LadderDto -> LadderGame 으로 변경
