+++ src/main/java/nextstep/ladder/ErrorMessage.java
@@ -9,4 +9,5 @@ public class ErrorMessage {
     public static final String INVALID_DIRECTION = "Invalid Direction";
     public static final String INVALID_LADDER_HEIGHT = "Invalid Ladder Height";
     public static final String INVALID_LADDER_GOAL = "Invalid Ladder Goal";
+    public static final String INVALID_LADDER_GAME = "Invalid Ladder Game";
 }
